work in progress

+++ instruction/Instruction.h
@@ -17,6 +17,7 @@
 
 #include "InstructionFactory.h"
 #include "../InstructionPointer.h"
+#include "../Data8.h"
 
 class InstructionFactory;
 class FlashMemory;
@@ -72,7 +73,7 @@ public:
             cycleCounter=6;
             uint opcode = flashMemory[IP];
             uint opcode1 = flashMemory[IP+1];
-            std::cout  << IP <<"  " << std::hex << opcode << " " << opcode1  << " [" << (int)xdata[4]->getValue() << "] ";;
+            std::cout  << IP <<"  " << std::hex << opcode << " " << opcode1  << " [" << xdata[4]->getValue() << "] ";;
             execution();
         }
         return instructionFactory.decode(flashMemory[IP]);
@@ -98,7 +99,7 @@ public:
             cycleCounter=4;
             uint opcode = flashMemory[IP];
             uint opcode1 = flashMemory[IP+1];
-            std::cout  << IP <<"  " << std::hex << opcode << " " << opcode1  << " [" << (int)xdata[4]->getValue() << "] ";;
+            std::cout  << IP <<"  " << std::hex << opcode << " " << opcode1  << " [" << xdata[4]->getValue() << "] ";;
             execution();
         }
         return instructionFactory.decode(flashMemory[IP]);
@@ -124,7 +125,7 @@ public:
             cycleCounter=3;
             uint opcode = flashMemory[IP];
             uint opcode1 = flashMemory[IP+1];
-            std::cout  << IP <<"  " << std::hex << opcode << " " << opcode1  << " [" << (int)xdata[4]->getValue() << "] ";;
+            std::cout  << IP <<"  " << std::hex << opcode << " " << opcode1  << " [" <<  xdata[4]->getValue() << "] ";;
             execution();
         }
         return instructionFactory.decode(flashMemory[IP]);
@@ -150,7 +151,7 @@ public:
             cycleCounter=2;
             uint opcode = flashMemory[IP];
             uint opcode1 = flashMemory[IP+1];
-            std::cout  << IP <<"  " << std::hex << opcode << " " << opcode1  << " [" << (int)xdata[4]->getValue() << "] ";;
+            std::cout  << IP <<"  " << std::hex << opcode << " " << opcode1  << " [" << xdata[4]->getValue() << "] ";;
             execution();
         }
         return instructionFactory.decode(flashMemory[IP]);
@@ -175,7 +176,7 @@ public:
         } else {
             cycleCounter=1;
             uint opcode = flashMemory[IP];
-            std::cout  << IP <<"  " << std::hex << opcode << " " << " [" << (int)xdata[4]->getValue() << "] ";
+            std::cout  << IP <<"  " << std::hex << opcode << " " << " [" << xdata[4]->getValue() << "] ";
             execution();
         }
         return instructionFactory.decode(flashMemory[IP]);
